% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_downscaling}
\alias{evaluate_downscaling}
\title{Evaluate downscaling model}
\usage{
evaluate_downscaling(
  model,
  test_input,
  test_target,
  metrics = c("rmse", "mae", "correlation"),
  plot = TRUE
)
}
\arguments{
\item{model}{The downscaling model (either BCSD or UNet)}

\item{test_input}{Array. Test input data.}

\item{test_target}{Array. Test target data.}

\item{metrics}{Character vector. Metrics to compute. Default: c("rmse", "mae", "correlation")}

\item{plot}{Logical. Whether to plot example predictions. Default: TRUE.}
}
\value{
List of evaluation metrics and potentially plots
}
\description{
Evaluate downscaling model
}
