% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srdrn.R
\name{predict.srdrn}
\alias{predict.srdrn}
\title{Predict method for SRDRN}
\usage{
\method{predict}{srdrn}(object, newdata, time_points = NULL)
}
\arguments{
\item{object}{A trained SRDRN object.}

\item{newdata}{A 3D array of shape (N_1, N_1, n) representing the new data to be predicted.}
}
\value{
A 3D array of shape (N_2, N_2, n) representing the predicted data.
}
\description{
This function makes predictions using a trained SRDRN model.
It takes a trained SRDRN object and new data as input,
normalizes the new data, and uses the model to make predictions.
The predictions are then rescaled back to the original range.
}
\details{
The predict method for the SRDRN class takes a trained SRDRN object and new data as input.
It normalizes the new data using the same min-max scaling used during training.
The new data is reshaped to match the input shape of the model,
and the model is used to make predictions.
The predictions are then rescaled back to the original range using the min-max scaling parameters
obtained during training.
The output is a 3D array of the predicted data.
}
\examples{
\dontrun{
# Generate dummy low-resolution (16×16) and high-resolution (32×32) data
n <- 20
input  <- array(runif(16 * 16 * n),  dim = c(16, 16, n))
target <- array(runif(32 * 32 * n),  dim = c(32, 32, n))

# Example 2: Spatio-temporal downscaling using time points
time_vec <- 1:n
model <- srdrn(
  input_data  = input,
  target_data = target,
  time_points = time_vec,
  cyclical_period = 365,
  temporal_layers = c(32, 64),
  epochs = 2,
  batch_size = 4
)

n_new <- 3
newdata <- array(runif(16 * 16 * n_new),
                     dim = c(16, 16, n_new))
predictions <- predict(model, newdata, 1:n_new)
}

}
