% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_italy.R
\docType{data}
\name{weather_italy}
\alias{weather_italy}
\title{Daily Weather Data for Italy}
\format{
A 4-dimensional array with dimensions:
\describe{
\item{longitude}{Longitude grid points (0.2 degree resolution)}
\item{latitude}{Latitude grid points (0.2 degree resolution)}
\item{variable}{Weather variables: relative humidity, temperature, total precipitation}
\item{time}{Daily measurements from 2023-11-01 to 2023-12-31}
}
}
\usage{
weather_italy
}
\description{
A dataset containing daily gridded weather variables for Italy.
}
\keyword{datasets}
